requirements.txt
boto3==1.28.*
aws-cdk-lib==2.96.0
constructs>=10.0.0,<11.0.0
